remove async crates sp1 prover
