+++ prover/Cargo.toml
@@ -16,36 +16,22 @@ p3-challenger = { workspace = true }
 p3-baby-bear = { workspace = true }
 p3-bn254-fr = { workspace = true }
 p3-commit = { workspace = true }
-p3-util = { workspace = true }
 bincode = "1.3.3"
 serde = { version = "1.0", features = ["derive", "rc"] }
-backtrace = "0.3.71"
 rayon = "1.10.0"
 itertools = "0.13.0"
 tracing = "0.1.40"
 tracing-subscriber = "0.3.18"
 serde_json = "1.0.120"
 clap = { version = "4.5.8", features = ["derive", "env"] }
-sha2 = "0.10.8"
 hex = "0.4.3"
 anyhow = "1.0.83"
-size = "0.4.1"
 dirs = "5.0.1"
 tempfile = "3.10.1"
-tokio = { version = "1.38.0", features = ["full"] }
-reqwest = { version = "0.12.4", features = [
-  "rustls-tls",
-  "trust-dns",
-  "stream",
-] }
-indicatif = "0.17.8"
-futures = "0.3.30"
 subtle-encoding = "0.5.1"
 serial_test = "3.1.1"
 num-bigint = "0.4.6"
 thiserror = "1.0.60"
-bytemuck = "1.16.0"
-rand = "0.8.4"
 
 [[bin]]
 name = "build_plonk_bn254"
