+++ prover/src/utils.rs
@@ -3,7 +3,6 @@ use std::{
     io::Read,
 };
 
-use futures::Future;
 use p3_baby_bear::BabyBear;
 use p3_bn254_fr::Bn254Fr;
 use p3_field::AbstractField;
@@ -14,7 +13,6 @@ use sp1_core::{
     runtime::{Program, Runtime},
     utils::SP1CoreOpts,
 };
-use tokio::{runtime, task::block_in_place};
 
 use crate::SP1CoreProofData;
 
@@ -84,17 +82,4 @@ pub fn words_to_bytes_be(words: &[u32; 8]) -> [u8; 32] {
         bytes[i * 4..(i + 1) * 4].copy_from_slice(&word_bytes);
     }
     bytes
-}
-
-/// Utility method for blocking on an async function. If we're already in a tokio runtime, we'll
-/// block in place. Otherwise, we'll create a new runtime.
-pub fn block_on<T>(fut: impl Future<Output = T>) -> T {
-    // Handle case if we're already in an tokio runtime.
-    if let Ok(handle) = runtime::Handle::try_current() {
-        block_in_place(|| handle.block_on(fut))
-    } else {
-        // Otherwise create a new runtime.
-        let rt = runtime::Runtime::new().expect("Failed to create a new runtime");
-        rt.block_on(fut)
-    }
 }
