+++ sdk/src/network/prover.rs
@@ -1,5 +1,6 @@
 use std::{env, time::Duration};
 
+use crate::install::block_on;
 use crate::proto::network::ProofMode;
 use crate::{
     network::client::{NetworkClient, DEFAULT_PROVER_NETWORK_RPC},
@@ -10,7 +11,6 @@ use crate::{SP1Context, SP1ProofKind, SP1ProofWithPublicValues, SP1ProvingKey, S
 use anyhow::Result;
 use serde::de::DeserializeOwned;
 use sp1_core::utils::SP1ProverOpts;
-use sp1_prover::utils::block_on;
 use sp1_prover::{SP1Prover, SP1Stdin, SP1_CIRCUIT_VERSION};
 use tokio::time::sleep;
