+++ sdk/src/provers/local.rs
@@ -4,7 +4,8 @@ use sp1_prover::{SP1Prover, SP1Stdin};
 use sysinfo::System;
 
 use crate::{
-    Prover, SP1Proof, SP1ProofKind, SP1ProofWithPublicValues, SP1ProvingKey, SP1VerifyingKey,
+    install::try_install_plonk_bn254_artifacts, Prover, SP1Proof, SP1ProofKind,
+    SP1ProofWithPublicValues, SP1ProvingKey, SP1VerifyingKey,
 };
 
 use super::ProverType;
@@ -79,7 +80,7 @@ impl Prover for LocalProver {
                 &outer_proof.proof,
             )
         } else {
-            sp1_prover::build::try_install_plonk_bn254_artifacts()
+            try_install_plonk_bn254_artifacts()
         };
         let proof = self
             .prover
